feat: TET-849 replace dot notation with a single component
