+++ src/components/Card/Card.tsx
@@ -1,28 +1,36 @@
-import type { CardComponent } from './Card.props';
+import { FC } from 'react';
+
+import type { CardProps } from './Card.props';
 import { cardConfig } from './Card.styles';
-import { CardContent } from './CardContent/CardContent';
-import { CardFooter } from './CardFooter';
+import { CardFooter } from './CardFooter/CardFooter';
 import { CardHeader } from './CardHeader/CardHeader';
 
 import { mergeConfigWithCustom } from '@/services';
 import { tet } from '@/tetrisly';
 
-export const Card: CardComponent = ({ custom, children, ...rest }) => {
-  const styles = mergeConfigWithCustom({
-    defaultConfig: cardConfig,
-    custom,
-  });
+export const Card: FC<CardProps> = ({
+  custom,
+  children,
+  header,
+  footer,
+  ...rest
+}) => {
+  const styles = getStylesProps(custom);
   return (
-    <tet.div data-testid="card" {...styles} {...rest}>
-      {children}
+    <tet.div data-testid="card" {...styles.container} {...rest}>
+      {header && <CardHeader styles={styles.header} {...header} />}
+      <tet.div data-testid="card-content" {...styles.content}>
+        {children}
+      </tet.div>
+      {footer && <CardFooter styles={styles.footer} {...footer} />}
     </tet.div>
   );
 };
 
-Card.Header = CardHeader;
-Card.Content = CardContent;
-Card.Footer = CardFooter;
-
-Card.Header.displayName = 'Card.Header';
-Card.Content.displayName = 'Card.Content';
-Card.Footer.displayName = 'Card.Footer';
+function getStylesProps(custom: CardProps['custom']) {
+  const { innerElements, ...container } = mergeConfigWithCustom({
+    defaultConfig: cardConfig,
+    custom,
+  });
+  return { ...innerElements, container };
+}
