--- src/components/Card/CardContent/CardContent.props.ts
@@ -1,6 +1,0 @@
-import { CardContentConfig } from './CardContent.styles';
-
-export type CardContentProps = {
-  children: React.ReactNode;
-  custom?: CardContentConfig;
-};
