--- src/components/Card/CardContent/index.ts
@@ -1,3 +1,0 @@
-export * from './CardContent';
-export * from './CardContent.props';
-export * from './CardContent.styles';
