+++ src/components/Card/CardFooter/CardFooter.props.ts
@@ -1,8 +1,8 @@
-import type { CardFooterConfig } from './CardFooter.styles';
+import type { CardFooterConfig } from '../Card.styles';
 
 import type { GhostButtonProps } from '@/components/Button/Button.props';
 
 export type CardFooterProps = {
-  custom?: CardFooterConfig;
-  actions?: GhostButtonProps[];
+  styles: CardFooterConfig;
+  actions: GhostButtonProps[];
 };
