+++ src/components/Card/CardFooter/CardFooter.tsx
@@ -1,19 +1,17 @@
 import type { FC } from 'react';
 
 import type { CardFooterProps } from './CardFooter.props';
-import { cardFooterConfig } from './CardFooter.styles';
 
 import { Button } from '@/components/Button';
-import { mergeConfigWithCustom } from '@/services';
 import { tet } from '@/tetrisly';
 import { BaseProps } from '@/types';
 
 export const CardFooter: FC<CardFooterProps> = ({
-  custom,
+  styles: rawStyles,
   actions,
   ...rest
 }) => {
-  const styles = getStylesProps(custom);
+  const styles = getStylesProps(rawStyles);
   return (
     <tet.div data-testid="card-footer" {...styles.container} {...rest}>
       <Actions styles={styles.actions} actions={actions} />
@@ -23,24 +21,19 @@ export const CardFooter: FC<CardFooterProps> = ({
 
 type CardFooterActionProps = {
   actions: CardFooterProps['actions'];
-  styles: BaseProps;
+  styles: BaseProps | undefined;
 };
 
-const Actions: FC<CardFooterActionProps> = ({ actions, styles }) => {
-  if (!actions) return null;
-  return (
-    <tet.div data-testid="card-footer-actions" {...styles}>
-      {actions.map((action) => (
-        <Button key={action.label} size="small" variant="ghost" {...action} />
-      ))}
-    </tet.div>
-  );
-};
+const Actions: FC<CardFooterActionProps> = ({ actions, styles }) => (
+  <tet.div data-testid="card-footer-actions" {...styles}>
+    {actions.map((action) => (
+      <Button key={action.label} size="small" variant="ghost" {...action} />
+    ))}
+  </tet.div>
+);
 
-function getStylesProps(custom: CardFooterProps['custom']) {
-  const { innerElements, ...rest } = mergeConfigWithCustom({
-    defaultConfig: cardFooterConfig,
-    custom,
-  });
-  return { ...innerElements, container: rest };
+function getStylesProps(styles: CardFooterProps['styles']) {
+  const { actions, ...container } = styles;
+
+  return { actions, container };
 }
