--- src/components/Card/CardFooter/index.ts
@@ -1,3 +1,0 @@
-export * from './CardFooter';
-export * from './CardFooter.props';
-export * from './CardFooter.styles';
