+++ src/components/Card/CardHeader/CardHeader.props.ts
@@ -1,6 +1,6 @@
 import type { IconProps } from '@virtuslab/tetrisly-icons';
 
-import type { CardHeaderConfig } from './CardHeader.styles';
+import type { CardHeaderConfig } from '../Card.styles';
 
 import type { AvatarProps } from '@/components/Avatar';
 import type { BareButtonProps } from '@/components/Button/Button.props';
@@ -9,10 +9,9 @@ export type CardHeaderProps = {
   title: string;
   description?: string;
   beforeComponent?: BeforeComponentProps;
-  custom?: CardHeaderConfig;
+  styles: CardHeaderConfig;
   actions?: BareButtonProps[];
 };
-
 type BeforeComponentProps =
   | {
       icon: IconProps;
