--- src/components/Card/CardHeader/index.ts
@@ -1,3 +1,0 @@
-export * from './CardHeader';
-export * from './CardHeader.props';
-export * from './CardHeader.styles';
