+++ src/components/Card/index.ts
@@ -1,4 +1,4 @@
 export * from './Card';
 export * from './Card.props';
 export * from './Card.styles';
-export * from './CardHeader';
+export * from './CardHeader/CardHeader';
